Fix #3
Added an array in the class which will contain all the titles of the
Fees which have been added by our plugin in the cart.

Also attached a function to the WC Subscriptions hook to ensure the fees
our included in the Recurring totals displayed in the cart table on the
Cart & Checkout pages.

So, no changes need to be made in the renewal order genertaion process
as the fees have been included in the subscription price itself.

Also, changed the priority of the function which adds the fees to ensure
it is performed after WC Subscriptions so fees are calculated correctly.