fix #185
The PG fees are not removed from the Edit Order Page. I changed the woocommerce_process_shop_order_meta hook and added the woocommerce_before_save_order_items hook to update the order with the updated changes. For testing, check that all updated PG fees are saved in order, and also check that they are updated in the frontend order.